+++ src/lib/provable/int.ts
@@ -10,6 +10,13 @@ import type { Gadgets } from './gadgets/gadgets.js';
 import { withMessage } from './field.js';
 import { FieldVar } from './core/fieldvar.js';
 import { CircuitValue, prop } from './types/circuit-value.js';
+import {
+  assertLessThanGeneric,
+  assertLessThanOrEqualGeneric,
+  lessThanGeneric,
+  lessThanOrEqualGeneric,
+} from './gadgets/comparison.js';
+import { assert } from '../util/assert.js';
 
 // external API
 export { UInt8, UInt32, UInt64, Int64, Sign };
@@ -401,24 +408,10 @@ class UInt64 extends CircuitValue {
   lessThanOrEqual(y: UInt64) {
     if (this.value.isConstant() && y.value.isConstant()) {
       return Bool(this.value.toBigInt() <= y.value.toBigInt());
-    } else {
-      let xMinusY = this.value.sub(y.value).seal();
-      let yMinusX = xMinusY.neg();
-
-      let xMinusYFits = RangeCheck.isDefinitelyInRangeN(
-        UInt64.NUM_BITS,
-        xMinusY
-      );
-
-      let yMinusXFits = RangeCheck.isDefinitelyInRangeN(
-        UInt64.NUM_BITS,
-        yMinusX
-      );
-
-      xMinusYFits.or(yMinusXFits).assertEquals(true);
-      // x <= y if y - x fits in 64 bits
-      return yMinusXFits;
     }
+    return lessThanOrEqualGeneric(this.value, y.value, 1n << 64n, (v) =>
+      RangeCheck.rangeCheckN(UInt64.NUM_BITS, v)
+    );
   }
 
   /**
@@ -426,16 +419,15 @@ class UInt64 extends CircuitValue {
    */
   assertLessThanOrEqual(y: UInt64, message?: string) {
     if (this.value.isConstant() && y.value.isConstant()) {
-      let x0 = this.value.toBigInt();
-      let y0 = y.value.toBigInt();
-      if (x0 > y0) {
-        if (message !== undefined) throw Error(message);
-        throw Error(`UInt64.assertLessThanOrEqual: expected ${x0} <= ${y0}`);
-      }
-      return;
+      let [x0, y0] = [this.value.toBigInt(), y.value.toBigInt()];
+      return assert(
+        x0 <= y0,
+        message ?? `UInt64.assertLessThanOrEqual: expected ${x0} <= ${y0}`
+      );
     }
-    let yMinusX = y.value.sub(this.value).seal();
-    RangeCheck.rangeCheckN(UInt64.NUM_BITS, yMinusX, message);
+    assertLessThanOrEqualGeneric(this.value, y.value, (v) =>
+      RangeCheck.rangeCheckN(UInt64.NUM_BITS, v, message)
+    );
   }
 
   /**
@@ -443,14 +435,28 @@ class UInt64 extends CircuitValue {
    * Checks if a {@link UInt64} is less than another one.
    */
   lessThan(y: UInt64) {
-    return this.lessThanOrEqual(y).and(this.value.equals(y.value).not());
+    if (this.value.isConstant() && y.value.isConstant()) {
+      return Bool(this.value.toBigInt() < y.value.toBigInt());
+    }
+    return lessThanGeneric(this.value, y.value, 1n << 64n, (v) =>
+      RangeCheck.rangeCheckN(UInt64.NUM_BITS, v)
+    );
   }
 
   /**
    * Asserts that a {@link UInt64} is less than another one.
    */
   assertLessThan(y: UInt64, message?: string) {
-    this.lessThan(y).assertEquals(true, message);
+    if (this.value.isConstant() && y.value.isConstant()) {
+      let [x0, y0] = [this.value.toBigInt(), y.value.toBigInt()];
+      return assert(
+        x0 < y0,
+        message ?? `UInt64.assertLessThan: expected ${x0} < ${y0}`
+      );
+    }
+    assertLessThanGeneric(this.value, y.value, (v) =>
+      RangeCheck.rangeCheckN(UInt64.NUM_BITS, v, message)
+    );
   }
 
   /**
@@ -471,7 +477,7 @@ class UInt64 extends CircuitValue {
    * Checks if a {@link UInt64} is greater than or equal to another one.
    */
   greaterThanOrEqual(y: UInt64) {
-    return this.lessThan(y).not();
+    return y.lessThanOrEqual(this);
   }
 
   /**
@@ -852,21 +858,10 @@ class UInt32 extends CircuitValue {
   lessThanOrEqual(y: UInt32) {
     if (this.value.isConstant() && y.value.isConstant()) {
       return Bool(this.value.toBigInt() <= y.value.toBigInt());
-    } else {
-      let xMinusY = this.value.sub(y.value).seal();
-      let yMinusX = xMinusY.neg();
-      let xMinusYFits = RangeCheck.isDefinitelyInRangeN(
-        UInt32.NUM_BITS,
-        xMinusY
-      );
-      let yMinusXFits = RangeCheck.isDefinitelyInRangeN(
-        UInt32.NUM_BITS,
-        yMinusX
-      );
-      xMinusYFits.or(yMinusXFits).assertEquals(true);
-      // x <= y if y - x fits in 64 bits
-      return yMinusXFits;
     }
+    return lessThanOrEqualGeneric(this.value, y.value, 1n << 32n, (v) =>
+      RangeCheck.rangeCheckN(UInt32.NUM_BITS, v)
+    );
   }
 
   /**
@@ -874,30 +869,43 @@ class UInt32 extends CircuitValue {
    */
   assertLessThanOrEqual(y: UInt32, message?: string) {
     if (this.value.isConstant() && y.value.isConstant()) {
-      let x0 = this.value.toBigInt();
-      let y0 = y.value.toBigInt();
-      if (x0 > y0) {
-        if (message !== undefined) throw Error(message);
-        throw Error(`UInt32.assertLessThanOrEqual: expected ${x0} <= ${y0}`);
-      }
-      return;
+      let [x0, y0] = [this.value.toBigInt(), y.value.toBigInt()];
+      return assert(
+        x0 <= y0,
+        message ?? `UInt32.assertLessThanOrEqual: expected ${x0} <= ${y0}`
+      );
     }
-    let yMinusX = y.value.sub(this.value).seal();
-    RangeCheck.rangeCheckN(UInt32.NUM_BITS, yMinusX, message);
+    assertLessThanOrEqualGeneric(this.value, y.value, (v) =>
+      RangeCheck.rangeCheckN(UInt32.NUM_BITS, v, message)
+    );
   }
 
   /**
    * Checks if a {@link UInt32} is less than another one.
    */
   lessThan(y: UInt32) {
-    return this.lessThanOrEqual(y).and(this.value.equals(y.value).not());
+    if (this.value.isConstant() && y.value.isConstant()) {
+      return Bool(this.value.toBigInt() < y.value.toBigInt());
+    }
+    return lessThanGeneric(this.value, y.value, 1n << 64n, (v) =>
+      RangeCheck.rangeCheckN(UInt64.NUM_BITS, v)
+    );
   }
 
   /**
    * Asserts that a {@link UInt32} is less than another one.
    */
   assertLessThan(y: UInt32, message?: string) {
-    this.lessThan(y).assertEquals(true, message);
+    if (this.value.isConstant() && y.value.isConstant()) {
+      let [x0, y0] = [this.value.toBigInt(), y.value.toBigInt()];
+      return assert(
+        x0 < y0,
+        message ?? `UInt32.assertLessThan: expected ${x0} < ${y0}`
+      );
+    }
+    assertLessThanGeneric(this.value, y.value, (v) =>
+      RangeCheck.rangeCheckN(UInt32.NUM_BITS, v, message)
+    );
   }
 
   /**
@@ -918,7 +926,7 @@ class UInt32 extends CircuitValue {
    * Checks if a {@link UInt32} is greater than or equal to another one.
    */
   greaterThanOrEqual(y: UInt32) {
-    return this.lessThan(y).not();
+    return y.lessThanOrEqual(this);
   }
 
   /**
@@ -1348,7 +1356,12 @@ class UInt8 extends Struct({
     if (this.value.isConstant() && y_.value.isConstant()) {
       return Bool(this.toBigInt() <= y_.toBigInt());
     }
-    throw Error('Not implemented');
+    return lessThanOrEqualGeneric(
+      this.value,
+      y_.value,
+      1n << 8n,
+      RangeCheck.rangeCheck8
+    );
   }
 
   /**
@@ -1365,7 +1378,12 @@ class UInt8 extends Struct({
     if (this.value.isConstant() && y_.value.isConstant()) {
       return Bool(this.toBigInt() < y_.toBigInt());
     }
-    throw Error('Not implemented');
+    return lessThanGeneric(
+      this.value,
+      y_.value,
+      1n << 8n,
+      RangeCheck.rangeCheck8
+    );
   }
 
   /**
@@ -1379,17 +1397,18 @@ class UInt8 extends Struct({
   assertLessThan(y: UInt8 | bigint | number, message?: string) {
     let y_ = UInt8.from(y);
     if (this.value.isConstant() && y_.value.isConstant()) {
-      let x0 = this.toBigInt();
-      let y0 = y_.toBigInt();
-      if (x0 >= y0) {
-        if (message !== undefined) throw Error(message);
-        throw Error(`UInt8.assertLessThan: expected ${x0} < ${y0}`);
-      }
-      return;
+      let [x0, y0] = [this.value.toBigInt(), y_.value.toBigInt()];
+      return assert(
+        x0 < y0,
+        message ?? `UInt8.assertLessThan: expected ${x0} < ${y0}`
+      );
     }
-    // x < y  <=>  x + 1 <= y
-    let xPlus1 = new UInt8(this.value.add(1).value);
-    xPlus1.assertLessThanOrEqual(y, message);
+    try {
+      // 2^16 < p - 2^8, so we satisfy the assumption of `assertLessThanGeneric`
+      assertLessThanGeneric(this.value, y_.value, RangeCheck.rangeCheck16);
+    } catch (err) {
+      throw withMessage(err, message);
+    }
   }
 
   /**
@@ -1403,18 +1422,19 @@ class UInt8 extends Struct({
   assertLessThanOrEqual(y: UInt8 | bigint | number, message?: string) {
     let y_ = UInt8.from(y);
     if (this.value.isConstant() && y_.value.isConstant()) {
-      let x0 = this.toBigInt();
-      let y0 = y_.toBigInt();
-      if (x0 > y0) {
-        if (message !== undefined) throw Error(message);
-        throw Error(`UInt8.assertLessThanOrEqual: expected ${x0} <= ${y0}`);
-      }
-      return;
+      let [x0, y0] = [this.value.toBigInt(), y_.value.toBigInt()];
+      return assert(
+        x0 <= y0,
+        message ?? `UInt8.assertLessThanOrEqual: expected ${x0} <= ${y0}`
+      );
     }
     try {
-      // x <= y  <=>  y - x >= 0  which is implied by  y - x in [0, 2^16)
-      let yMinusX = y_.value.sub(this.value).seal();
-      RangeCheck.rangeCheck16(yMinusX);
+      // 2^16 < p - 2^8, so we satisfy the assumption of `assertLessThanOrEqualGeneric`
+      assertLessThanOrEqualGeneric(
+        this.value,
+        y_.value,
+        RangeCheck.rangeCheck16
+      );
     } catch (err) {
       throw withMessage(err, message);
     }